Monte Carlo
C++ version of the Monte Carlo program for calculating Pi using
pthreads.

+++ MonteCarlo/timer.h
@@ -5,26 +5,23 @@
 //  Created by James Williams on 9/23/15.
 //  Copyright © 2015 James Williams. All rights reserved.
 //
-
+#include <time.h>
 #ifndef timer_h
 #define timer_h
 
 
-class timer
-{
+class timer {
 private:
     unsigned long begTime;
-    
 public:
-    void start()
-    {
+    void start() {
         begTime = clock();
     }
     
-    unsigned long elapsedTime()
-    {
-        return ((unsigned long) clock() - begTime);
+    unsigned long elapsedTime() {
+        return (clock() - begTime);
     }
 };
+
 
 #endif /* timer_h */
